Fix parsing of unneccesary long IDs
Preserve the length of the ID when parsing, so that we can reproduce
the exact same stream again. Especially for EMV applications which
do not always strictly adhere to the TLV standard this is important.
(so that e.g. an ID 9F07 is not shrunk to 87 which would the same tag
from a TLV spec standpoint)