Fix build/loading of resource files
When creating a File object with a resource,
the resource has to be in a directory on disk
and may not be in an archive.
Converting to a File is the old way of handling this.
Changed to an InputStream, these work in any case.

Additionally added resources to the pom files,
otherwise they where not included in the maven builds
leading to a broken update site.